pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-automated-trading"
version = "1.0.0"
description = "AI-powered automated stock trading system with ethical filtering"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "yfinance>=0.2.18",
    "pandas>=2.1.0",
    "numpy>=1.25.0,<2.3.0",
    "openai>=1.3.0",
    "textblob>=0.17.1",
    "alpaca-py>=0.40.0",
    "python-telegram-bot>=20.6",
    "schedule>=1.2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "reportlab>=4.0.0",
    "websockets>=10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 